refactoring: 서비스에 있는  business method -> Entity로 이동

+++ src/main/java/com/example/theater_proj/movie/dto/request/ReservationRequest.java
@@ -6,8 +6,8 @@ import com.example.theater_proj.movie.model.PaymentStatus;
 import java.util.List;
 
 public record ReservationRequest(
-        int screeningId,
-        List<Integer> seatsIds,
+        Long screeningId,
+        List<Long> seatsIds,
         int price
 ) {
